visualize/requirements.txt
matplotlib
numpy
torch
torchvision
Pillow
tensorboard
ipykernel
pyyaml